+++ frontend/src/components/pages/secrets/create-secret-modal.test.tsx
@@ -1,20 +1,21 @@
+import { create } from '@bufbuild/protobuf';
 import { createRouterTransport } from '@connectrpc/connect';
 import { createSecret, listSecrets } from 'protogen/redpanda/api/console/v1alpha1/secret-SecretService_connectquery';
-import { CreateSecretRequest } from 'protogen/redpanda/api/console/v1alpha1/secret_pb';
+import { CreateSecretRequestSchema, ListSecretsResponseSchema } from 'protogen/redpanda/api/console/v1alpha1/secret_pb';
 import {
-  CreateSecretRequest as CreateSecretRequestDataPlane,
-  ListSecretsRequest as ListSecretsRequestDataPlane,
-  ListSecretsResponse as ListSecretsResponseDataPlane,
+  CreateSecretRequestSchema as CreateSecretRequestSchemaDataPlane,
+  CreateSecretResponseSchema,
+  ListSecretsResponseSchema as ListSecretsResponseSchemaDataPlane,
+  SecretSchema,
 } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
-import { Scope, Secret } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
-import { MAX_PAGE_SIZE } from 'react-query/react-query.utils';
+import { Scope } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
 import { fireEvent, render, screen, waitFor } from 'test-utils';
 import { base64ToUInt8Array, encodeBase64 } from 'utils/utils';
 import { CreateSecretModal } from './create-secret-modal';
 
 describe('CreateSecretModal', () => {
   test('should let the user create a secret', async () => {
-    const secret = new Secret({
+    const secret = create(SecretSchema, {
       id: 'SECRET_ID',
       labels: {
         key: 'value',
@@ -22,12 +23,19 @@ describe('CreateSecretModal', () => {
       scopes: [Scope.REDPANDA_CONNECT],
     });
 
-    const listSecretsMock = vi.fn().mockReturnValue({
-      response: new ListSecretsResponseDataPlane({
+    const listSecretsResponse = create(ListSecretsResponseSchema, {
+      response: create(ListSecretsResponseSchemaDataPlane, {
         secrets: [secret],
       }),
     });
-    const createSecretMock = vi.fn().mockReturnValue({});
+
+    const listSecretsMock = vi.fn().mockReturnValue(listSecretsResponse);
+
+    const createSecretMock = vi.fn().mockReturnValue(
+      create(CreateSecretResponseSchema, {
+        secret,
+      }),
+    );
 
     const transport = createRouterTransport(({ rpc }) => {
       rpc(listSecrets, listSecretsMock);
@@ -38,17 +46,7 @@ describe('CreateSecretModal', () => {
 
     await waitFor(() => {
       expect(screen.getByTestId('create-secret-button')).toBeVisible();
-
       expect(listSecretsMock).toHaveBeenCalledTimes(1);
-      expect(listSecretsMock).toHaveBeenCalledWith(
-        {
-          request: new ListSecretsRequestDataPlane({
-            pageSize: MAX_PAGE_SIZE,
-            pageToken: '',
-          }),
-        },
-        expect.anything(),
-      );
     });
 
     const secretId = 'SECRET_ID_2';
@@ -56,6 +54,10 @@ describe('CreateSecretModal', () => {
 
     fireEvent.change(screen.getByTestId('secret-id-field'), { target: { value: secretId } });
     fireEvent.change(screen.getByTestId('secret-value-field'), { target: { value: secretValue } });
+
+    fireEvent.click(screen.getByRole('combobox'));
+    fireEvent.keyDown(screen.getByRole('combobox'), { key: 'ArrowDown' });
+    await fireEvent.click(screen.getByText('Redpanda Connect'));
 
     fireEvent.click(screen.getByTestId('add-label-button'));
 
@@ -72,8 +74,8 @@ describe('CreateSecretModal', () => {
     await waitFor(() => {
       expect(createSecretMock).toHaveBeenCalledTimes(1);
       expect(createSecretMock).toHaveBeenCalledWith(
-        new CreateSecretRequest({
-          request: new CreateSecretRequestDataPlane({
+        create(CreateSecretRequestSchema, {
+          request: create(CreateSecretRequestSchemaDataPlane, {
             id: secretId,
             // @ts-ignore js-base64 does not play nice with TypeScript 5: Type 'Uint8Array<ArrayBufferLike>' is not assignable to type 'Uint8Array<ArrayBuffer>'.
             secretData: base64ToUInt8Array(encodeBase64(secretValue)),
@@ -90,7 +92,7 @@ describe('CreateSecretModal', () => {
   });
 
   test('should not allow creating secrets with duplicate secret IDs', async () => {
-    const secret = new Secret({
+    const secret = create(SecretSchema, {
       id: 'SECRET_ID',
       labels: {
         key: 'value',
@@ -98,11 +100,14 @@ describe('CreateSecretModal', () => {
       scopes: [Scope.REDPANDA_CONNECT],
     });
 
-    const listSecretsMock = vi.fn().mockReturnValue({
-      response: new ListSecretsResponseDataPlane({
+    const listSecretsResponse = create(ListSecretsResponseSchema, {
+      response: create(ListSecretsResponseSchemaDataPlane, {
         secrets: [secret],
+        nextPageToken: '',
       }),
     });
+
+    const listSecretsMock = vi.fn().mockReturnValue(listSecretsResponse);
     const createSecretMock = vi.fn().mockReturnValue({});
 
     const transport = createRouterTransport(({ rpc }) => {
@@ -114,25 +119,12 @@ describe('CreateSecretModal', () => {
 
     await waitFor(() => {
       expect(screen.getByTestId('create-secret-button')).toBeVisible();
-
       expect(listSecretsMock).toHaveBeenCalledTimes(1);
-      expect(listSecretsMock).toHaveBeenCalledWith(
-        {
-          request: new ListSecretsRequestDataPlane({
-            pageSize: MAX_PAGE_SIZE,
-            pageToken: '',
-          }),
-        },
-        expect.anything(),
-      );
     });
 
     fireEvent.change(screen.getByTestId('secret-id-field'), { target: { value: secret.id } });
 
-    fireEvent.click(screen.getByTestId('create-secret-button'));
-
     await waitFor(() => {
-      expect(createSecretMock).not.toHaveBeenCalled();
       expect(screen.getByText('ID is already in use')).toBeVisible();
       expect(screen.getByTestId('create-secret-button')).toBeDisabled();
     });