+++ frontend/src/components/pages/secrets/create-secret-modal.tsx
@@ -1,6 +1,9 @@
+import { create } from '@bufbuild/protobuf';
 import {
   Button,
   ButtonGroup,
+  FormErrorMessage,
+  FormField,
   Modal,
   ModalBody,
   ModalCloseButton,
@@ -8,12 +11,15 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
+  Select,
   Stack,
   Text,
+  UnorderedList,
+  isMultiValue,
 } from '@redpanda-data/ui';
 import { formOptions } from '@tanstack/react-form';
 import { useAppForm } from 'components/form/form';
-import { CreateSecretRequest, Scope } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
+import { CreateSecretRequestSchema, Scope } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
 import type { ReactNode } from 'react';
 import { useCreateSecretMutationWithToast, useListSecretsQuery } from 'react-query/api/secret';
 import { base64ToUInt8Array, encodeBase64 } from 'utils/utils';
@@ -30,6 +36,8 @@ interface CreateSecretModalProps {
 export const CreateSecretModal = ({ isOpen, onClose, customSecretSchema, helperText }: CreateSecretModalProps) => {
   const { data: secretList } = useListSecretsQuery();
 
+  console.log('CreateSecretModal secretList: ', secretList);
+
   // Secret creation mutation
   const { mutateAsync: createSecret, isPending: isCreateSecretPending } = useCreateSecretMutationWithToast();
 
@@ -40,12 +48,23 @@ export const CreateSecretModal = ({ isOpen, onClose, customSecretSchema, helperT
     form.reset();
   };
 
+  // Form type
+  interface Secret {
+    id: string;
+    value: string;
+    labels: string[];
+    scopes: Scope[];
+  }
+
+  const defaultValues: Secret = {
+    id: '',
+    value: '',
+    labels: [],
+    scopes: [],
+  };
+
   const formOpts = formOptions({
-    defaultValues: {
-      id: '',
-      value: '',
-      labels: [],
-    },
+    defaultValues: defaultValues,
     validators: {
       onChange: finalSchema,
     },
@@ -57,11 +76,11 @@ export const CreateSecretModal = ({ isOpen, onClose, customSecretSchema, helperT
         }
       }
 
-      const request = new CreateSecretRequest({
+      const request = create(CreateSecretRequestSchema, {
         id: value.id,
         // @ts-ignore js-base64 does not play nice with TypeScript 5: Type 'Uint8Array<ArrayBufferLike>' is not assignable to type 'Uint8Array<ArrayBuffer>'.
         secretData: base64ToUInt8Array(encodeBase64(value.value)),
-        scopes: [Scope.REDPANDA_CONNECT],
+        scopes: value.scopes || [],
         labels: labelsMap,
       });
 
@@ -108,6 +127,42 @@ export const CreateSecretModal = ({ isOpen, onClose, customSecretSchema, helperT
                     <field.PasswordField label="Value" data-testid="secret-value-field" helperText={helperText} />
                   )}
                 </form.AppField>
+                <form.AppField name="scopes">
+                  {({ state, handleChange, handleBlur }) => (
+                    <FormField label="Scopes" errorText=" " isInvalid={state.meta.errors?.length > 0}>
+                      <Select
+                        placeholder="Select scopes"
+                        data-testid="secret-scopes-field"
+                        onChange={(nextValue) => {
+                          if (isMultiValue(nextValue) && nextValue) {
+                            handleChange(nextValue.map(({ value }) => value));
+                          }
+                        }}
+                        options={[
+                          { label: 'Redpanda Connect', value: Scope.REDPANDA_CONNECT },
+                          { label: 'Redpanda Cluster', value: Scope.REDPANDA_CLUSTER },
+                        ]}
+                        isMulti
+                        onBlur={handleBlur}
+                      />
+                      {
+                        // Display error messages like tanstack/react-form fields.
+                        state?.meta.errors?.length > 0 && (
+                          <FormErrorMessage>
+                            <UnorderedList>
+                              {state.meta.errors?.map((error) => (
+                                <li key={error.path}>
+                                  <Text color="red.500">{error.message}</Text>
+                                </li>
+                              ))}
+                            </UnorderedList>
+                          </FormErrorMessage>
+                        )
+                      }
+                    </FormField>
+                  )}
+                </form.AppField>
+                {/* @ts-ignore - labels is a valid field name, @tanstack/form needs updating to infer deeply nested form field types */}
                 <form.AppField name="labels" mode="array">
                   {(field) => (
                     <field.KeyValueField