+++ frontend/src/react-query/api/pipeline.tsx
@@ -1,5 +1,6 @@
-import type { PartialMessage } from '@bufbuild/protobuf';
-import { useMutation, useQuery } from '@connectrpc/connect-query';
+import { create } from '@bufbuild/protobuf';
+import type { GenMessage } from '@bufbuild/protobuf/codegenv1';
+import { createConnectQueryKey, useMutation, useQuery } from '@connectrpc/connect-query';
 import { useQueryClient } from '@tanstack/react-query';
 import {
   createPipeline,
@@ -13,22 +14,26 @@ import {
   updatePipeline,
 } from 'protogen/redpanda/api/console/v1alpha1/pipeline-PipelineService_connectquery';
 import {
-  GetPipelineRequest,
-  GetPipelinesBySecretsRequest,
-  GetPipelinesForSecretRequest,
-  ListPipelinesRequest,
+  type GetPipelineRequest,
+  GetPipelineRequestSchema,
+  GetPipelinesBySecretsRequestSchema,
+  GetPipelinesForSecretRequestSchema,
+  type ListPipelinesRequest,
+  ListPipelinesRequestSchema,
   type ListPipelinesResponse,
+  PipelineService,
 } from 'protogen/redpanda/api/console/v1alpha1/pipeline_pb';
 import {
-  GetPipelineRequest as GetPipelineRequestDataPlane,
+  GetPipelineRequestSchema as GetPipelineRequestSchemaDataPlane,
   type GetPipelineResponse,
-  GetPipelinesBySecretsRequest as GetPipelinesBySecretsRequestDataPlane,
-  GetPipelinesForSecretRequest as GetPipelinesForSecretRequestDataPlane,
-  ListPipelinesRequest as ListPipelinesRequestDataPlane,
+  GetPipelinesBySecretsRequestSchema as GetPipelinesBySecretsRequestSchemaDataPlane,
+  GetPipelinesForSecretRequestSchema as GetPipelinesForSecretRequestSchemaDataPlane,
+  type ListPipelinesRequest as ListPipelinesRequestDataPlane,
+  ListPipelinesRequestSchema as ListPipelinesRequestSchemaDataPlane,
   type Pipeline,
 } from 'protogen/redpanda/api/dataplane/v1/pipeline_pb';
 import type { Secret } from 'protogen/redpanda/api/dataplane/v1/secret_pb';
-import { MAX_PAGE_SIZE, type QueryOptions } from 'react-query/react-query.utils';
+import { MAX_PAGE_SIZE, type MessageInit, type QueryOptions } from 'react-query/react-query.utils';
 import { useInfiniteQueryWithAllPages } from 'react-query/use-infinite-query-with-all-pages';
 import { TOASTS, formatToastErrorMessageGRPC, showToast } from 'utils/toast.utils';
 
@@ -38,14 +43,14 @@ export const REDPANDA_CONNECT_LOGS_TIME_WINDOW_HOURS = 5;
 
 export const useGetPipelineQuery = (
   { id }: { id: Pipeline['id'] },
-  options?: QueryOptions<GetPipelineRequest, GetPipelineResponse, GetPipelineResponse> & {
+  options?: QueryOptions<GenMessage<GetPipelineRequest>, GetPipelineResponse> & {
     refetchInterval?: number | false;
     refetchIntervalInBackground?: boolean;
     refetchOnWindowFocus?: 'always' | boolean;
   },
 ) => {
-  const getPipelineRequestDataPlane = new GetPipelineRequestDataPlane({ id });
-  const getPipelineRequest = new GetPipelineRequest({ request: getPipelineRequestDataPlane });
+  const getPipelineRequestDataPlane = create(GetPipelineRequestSchemaDataPlane, { id });
+  const getPipelineRequest = create(GetPipelineRequestSchema, { request: getPipelineRequestDataPlane });
   return useQuery(getPipeline, getPipelineRequest, {
     enabled: options?.enabled,
     refetchInterval: options?.refetchInterval,
@@ -55,10 +60,10 @@ export const useGetPipelineQuery = (
 };
 
 export const useListPipelinesQuery = (
-  input?: PartialMessage<ListPipelinesRequestDataPlane>,
-  options?: QueryOptions<ListPipelinesRequestDataPlane, ListPipelinesResponse, ListPipelinesResponse>,
+  input?: MessageInit<ListPipelinesRequestDataPlane>,
+  options?: QueryOptions<GenMessage<ListPipelinesRequest>, ListPipelinesResponse>,
 ) => {
-  const listPipelinesRequestDataPlane = new ListPipelinesRequestDataPlane({
+  const listPipelinesRequestDataPlane = create(ListPipelinesRequestSchemaDataPlane, {
     pageSize: MAX_PAGE_SIZE,
     pageToken: '',
     // TODO: Use once nameContains is not required anymore
@@ -72,13 +77,21 @@ export const useListPipelinesQuery = (
     ...input,
   });
 
-  const listPipelinesRequest = new ListPipelinesRequest({
+  const listPipelinesRequest = create(ListPipelinesRequestSchema, {
     request: listPipelinesRequestDataPlane,
-  }) as PartialMessage<ListPipelinesRequest> & Required<Pick<PartialMessage<ListPipelinesRequest>, 'request'>>;
+  }) as MessageInit<ListPipelinesRequest> & Required<Pick<MessageInit<ListPipelinesRequest>, 'request'>>;
 
   const listPipelinesResult = useInfiniteQueryWithAllPages(listPipelines, listPipelinesRequest, {
     pageParamKey: 'request',
     enabled: options?.enabled,
+    // Required because of protobuf v2 reflection - it does not accept foreign fields when nested under "request", so the format needs to be a dataplane schema
+    getNextPageParam: (lastPage) =>
+      lastPage?.response?.nextPageToken
+        ? {
+            ...listPipelinesRequestDataPlane,
+            pageToken: lastPage.response?.nextPageToken,
+          }
+        : undefined,
   });
 
   const allRetrievedPipelines = listPipelinesResult?.data?.pages?.flatMap(({ response }) => response?.pipelines);
@@ -101,7 +114,12 @@ export const useCreatePipelineMutationWithToast = () => {
 
   return useMutation(createPipeline, {
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [listPipelines.service.typeName] });
+      await queryClient.invalidateQueries({
+        queryKey: createConnectQueryKey({
+          schema: PipelineService.method.listPipelines,
+          cardinality: 'infinite',
+        }),
+      });
       showToast({
         id: TOASTS.PIPELINE.CREATE.SUCCESS,
         title: 'Pipeline created successfully',
@@ -123,7 +141,12 @@ export const useUpdatePipelineMutationWithToast = () => {
 
   return useMutation(updatePipeline, {
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [listPipelines.service.typeName] });
+      await queryClient.invalidateQueries({
+        queryKey: createConnectQueryKey({
+          schema: PipelineService.method.listPipelines,
+          cardinality: 'infinite',
+        }),
+      });
       showToast({
         id: TOASTS.PIPELINE.UPDATE.SUCCESS,
         title: 'Pipeline updated successfully',
@@ -145,7 +168,12 @@ export const useStartPipelineMutationWithToast = () => {
 
   return useMutation(startPipeline, {
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [listPipelines.service.typeName] });
+      await queryClient.invalidateQueries({
+        queryKey: createConnectQueryKey({
+          schema: PipelineService.method.listPipelines,
+          cardinality: 'infinite',
+        }),
+      });
       showToast({
         id: TOASTS.PIPELINE.START.SUCCESS,
         title: 'Pipeline started successfully',
@@ -167,7 +195,12 @@ export const useStopPipelineMutationWithToast = () => {
 
   return useMutation(stopPipeline, {
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [listPipelines.service.typeName] });
+      await queryClient.invalidateQueries({
+        queryKey: createConnectQueryKey({
+          schema: PipelineService.method.listPipelines,
+          cardinality: 'infinite',
+        }),
+      });
       showToast({
         id: TOASTS.PIPELINE.STOP.SUCCESS,
         title: 'Pipeline stopped successfully',
@@ -189,7 +222,12 @@ export const useDeletePipelineMutationWithToast = () => {
 
   return useMutation(deletePipeline, {
     onSuccess: async () => {
-      await queryClient.invalidateQueries({ queryKey: [listPipelines.service.typeName] });
+      await queryClient.invalidateQueries({
+        queryKey: createConnectQueryKey({
+          schema: PipelineService.method.listPipelines,
+          cardinality: 'infinite',
+        }),
+      });
       showToast({
         id: TOASTS.PIPELINE.DELETE.SUCCESS,
         title: 'Pipeline deleted successfully',
@@ -207,9 +245,9 @@ export const useDeletePipelineMutationWithToast = () => {
 };
 
 export const useGetPipelinesForSecretQuery = ({ secretId }: { secretId: Secret['id'] }) => {
-  const getPipelinesForSecretRequestDataPlane = new GetPipelinesForSecretRequestDataPlane({ secretId });
+  const getPipelinesForSecretRequestDataPlane = create(GetPipelinesForSecretRequestSchemaDataPlane);
 
-  const getPipelinesForSecretRequest = new GetPipelinesForSecretRequest({
+  const getPipelinesForSecretRequest = create(GetPipelinesForSecretRequestSchema, {
     request: getPipelinesForSecretRequestDataPlane,
   });
 
@@ -219,11 +257,9 @@ export const useGetPipelinesForSecretQuery = ({ secretId }: { secretId: Secret['
 };
 
 export const useGetPipelinesBySecretsQuery = () => {
-  const getPipelinesBySecretsRequestDataPlane = new GetPipelinesBySecretsRequestDataPlane({
-    secretId: '',
-  });
+  const getPipelinesBySecretsRequestDataPlane = create(GetPipelinesBySecretsRequestSchemaDataPlane);
 
-  const getPipelinesBySecretsRequest = new GetPipelinesBySecretsRequest({
+  const getPipelinesBySecretsRequest = create(GetPipelinesBySecretsRequestSchema, {
     request: getPipelinesBySecretsRequestDataPlane,
   });
   return useQuery(getPipelinesBySecrets, getPipelinesBySecretsRequest);